Remove Load Balancer Experiment
TeamCity cannot run behind a load balancer, because the makers of teamcity
do not provide a health check method for the server. And because it does not
have that, it will always fail health checks, and be continually marked
as unhealthy.

I only put it behind a load balancer in the first place to hide the
IP address of the ECS Task anyways. Wanting to make a more secure
environment.

Well, this is good enough to demonstrate how to use TeamCity in an
AWS environment.